+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/coding/MultipleVariableDeclarationsCheckTest.java
@@ -35,7 +35,9 @@ public class MultipleVariableDeclarationsCheckTest extends BaseCheckTestSupport 
     @Override
     protected String getPath(String filename) throws IOException {
         return super.getPath("checks" + File.separator
-                + "coding" + File.separator + filename);
+                + "coding" + File.separator
+                + "multiplevariabledeclarations" + File.separator
+                + filename);
     }
 
     @Test
